+++ Source/Modifiers/_modifierDictionary.js
@@ -1,10 +1,8 @@
 const { parseCount } = require("../../helpers");
 
 const modifierWhitelist = [
-	"absorb-darkness.js",
 	"absorb-earth.js",
 	"absorb-fire.js",
-	"absorb-light.js",
 	"absorb-water.js",
 	"absorb-wind.js",
 	"curse-of-midas.js",