+++ Source/equipment/bloodaegis-sweeping.js
@@ -1,7 +1,7 @@
 const Equipment = require('../../Classes/Equipment.js');
 const { removeModifier, addBlock, dealDamage } = require('../combatantDAO.js');
 
-module.exports = new Equipment("Sweeping Blood Aegis", 2, "*Pay @{hpCost} hp to grant all allies @{block} block*\nCritical Hit: Block x@{critBonus}", "Darkness", effect, ["Charging Blood Aegis", "Heavy Blood Aegis"])
+module.exports = new Equipment("Sweeping Blood Aegis", 2, "*Pay @{hpCost} hp to grant all allies @{block} block*\nCritical Hit: Block x@{critBonus}", "Water", effect, ["Charging Blood Aegis", "Heavy Blood Aegis"])
 	.setCategory("Pact")
 	.setTargetingTags({ target: "all", team: "delver" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }])