+++ Source/equipment/inspiration-base.js
@@ -1,7 +1,7 @@
 const Equipment = require('../../Classes/Equipment.js');
 const { addModifier, removeModifier } = require('../combatantDAO.js');
 
-module.exports = new Equipment("Inspiration", 1, "Apply @{mod1Stacks} @{mod1} to an ally*\nCritical Hit: Apply @{mod2Stacks} @{mod2} to an delver", "Light", effect, ["Soothing Inspiration", "Sweeping Inspiration"])
+module.exports = new Equipment("Inspiration", 1, "Apply @{mod1Stacks} @{mod1} to an ally*\nCritical Hit: Apply @{mod2Stacks} @{mod2} to an delver", "Wind", effect, ["Soothing Inspiration", "Sweeping Inspiration"])
 	.setCategory("Spell")
 	.setTargetingTags({ target: "single", team: "delver" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }, { name: "Power Up", stacks: 50 }])