+++ Source/equipment/lifedrain-urgent.js
@@ -1,7 +1,7 @@
 const Equipment = require('../../Classes/Equipment.js');
 const { addModifier, dealDamage, gainHealth } = require('../combatantDAO.js');
 
-module.exports = new Equipment("Urgent Life Drain", 2, "*Strike a foe for @{damage} @{element} damage, then gain @{healing} hp (+@{speedBonus} speed bonus)*\nCritical Hit: Healing x@{critBonus}", "Darkness", effect, ["Flanking Life Drain", "Reactive Life Drain"])
+module.exports = new Equipment("Urgent Life Drain", 2, "*Strike a foe for @{damage} @{element} damage, then gain @{healing} hp (+@{speedBonus} speed bonus)*\nCritical Hit: Healing x@{critBonus}", "Water", effect, ["Flanking Life Drain", "Reactive Life Drain"])
 	.setCategory("Spell")
 	.setTargetingTags({ target: "single", team: "enemy" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }])