+++ Source/equipment/potion-watery.js
@@ -1,7 +1,7 @@
 const Equipment = require('../../Classes/Equipment.js');
 const { removeModifier, addModifier } = require('../combatantDAO.js');
 
-module.exports = new Equipment("Watery Potion", 2, "*Apply @{mod1Stacks} @{mod1} to a Water element combatant, or @{mod2Stacks} @{mod2} to someone else*\nCritical Hit: @{mod1}/@{mod2} x@{critBonus}", "Water", effect, ["Earthen Potion", "Inky Potion"])
+module.exports = new Equipment("Watery Potion", 2, "*Apply @{mod1Stacks} @{mod1} to a Water element combatant, or @{mod2Stacks} @{mod2} to someone else*\nCritical Hit: @{mod1}/@{mod2} x@{critBonus}", "Water", effect, ["Earthen Potion", "Windy Potion"])
 	.setCategory("Trinket")
 	.setTargetingTags({ target: "single", team: "any" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Regen", stacks: 4 }, { name: "Poison", stacks: 5 }, { name: "Regen", stacks: 8 }, { name: "Poison", stacks: 10 }])