+++ ChangeLog.md
@@ -5,6 +5,7 @@
    2. Vulnerabilties - Critical Hits and Elements for each combatant
    3. Intents - Targets and next two Move Names for enemies
    4. Health - HP and Modifiers for each combatant
+- Removed Darkness and Light elements; enemies, weapons, and archetypes have been re-elemented
 - Fixed `/invite`
 - Slash commands can now be used from DMs (as long as they make sense)
 - New slash command: `/armory` that allows checking stats on a weapon given by name